fix(rbac-registration): move cert_or_pk to its own file
