+++ rust/rbac-registration/src/cardano/cip509/types/mod.rs
@@ -1,17 +1,19 @@
 //! Types use in CIP-509
 
 pub use cert_key_hash::CertKeyHash;
+pub use cert_or_pk::CertOrPk;
 pub use key_local_ref::{KeyLocalRef, LocalRefInt};
 pub use payment_history::{Payment, PaymentHistory};
 pub use point_data::PointData;
 pub use point_tx_idx::PointTxnIdx;
 pub use role_data::RoleData;
-pub use role_data_record::{CertOrPk, RoleDataRecord};
+pub use role_data_record::RoleDataRecord;
 pub use role_number::RoleNumber;
 pub use tx_input_hash::TxInputHash;
 pub use validation_signature::ValidationSignature;
 
 mod cert_key_hash;
+mod cert_or_pk;
 mod key_local_ref;
 mod payment_history;
 mod point_data;
